0015-criar-service-e-controller-de-cidade

+++ src/main/java/com/estudo/myfoodapi/infrastructure/repository/CidadeRepositoryImpl.java
@@ -5,9 +5,11 @@ import com.estudo.myfoodapi.domain.repository.CidadeRepository;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import jakarta.transaction.Transactional;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
+import java.util.Objects;
 
 @Component
 public class CidadeRepositoryImpl implements CidadeRepository {
@@ -33,8 +35,11 @@ public class CidadeRepositoryImpl implements CidadeRepository {
 
     @Override
     @Transactional
-    public void remover(Cidade cidade) {
-        cidade = buscar(cidade.getId());
+    public void remover(Long id) {
+        Cidade cidade = buscar(id);
+        if (Objects.isNull(cidade)) {
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(cidade);
     }
 }
